切開action component
